requirements.txt
python-dotenv
requests
langchain
langgraph
langchain-google-genai
matplotlib
IPython